add extension test
