+++ rpc/client/rpc_test.go
@@ -63,8 +63,8 @@ func GetClients(t *testing.T, ns service.Service, conf *config.Config) []client.
 	require.NoError(t, err)
 
 	return []client.Client{
+    ncl,
 		getHTTPClient(t, conf),
-		ncl,
 	}
 }
